% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CryptoData.R
\name{format_pair}
\alias{format_pair}
\title{Format Cryptocurrency and Currency Pair}
\usage{
format_pair(coin_abbreviation, currency_abbreviation)
}
\arguments{
\item{coin_abbreviation}{Abbreviation of cryptocurrency to import (ex. "btc").}

\item{currency_abbreviation}{Abbreviation of local currency (ex. "usd").}
}
\description{
Formats cryptocurrency and local currency pair to load into the "get_coin_data" package. An example of the output will resemble "BTC-USD".
}
\examples{
## Enter coin abbreviation and local currency abbreviation in quotes, seperated by a comma.
# format_pair("btc", "usd")
}
