% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CryptoData.R
\name{get_coin_data}
\alias{get_coin_data}
\title{Import Cryptocurrency Data}
\usage{
get_coin_data(pair)
}
\arguments{
\item{pair}{Cryptocurrency and currency pair to import. Can be obtained with "format_pair" function.}
}
\description{
Imports open, close, high, and low price; as well as volume of selected cryptocurrency.
}
\examples{
## Type the cryptocurrency-local currency pair in quotes.
# get_coin_data("BTC-USD")

## Use the format_pair() function to get pair format.
# pair <- format_pair("btc", "usd")
# get_coin_data(pair)
}
