% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CryptoData.R
\name{low_52_week}
\alias{low_52_week}
\title{52-Week Lowest Price}
\usage{
low_52_week(coin_data)
}
\arguments{
\item{coin_data}{Dataframe produced by the "get_coin_data" function.}
}
\description{
Gets the lowest price within a 52 week period and prints the price and date of that occurrence.
}
\examples{
## Using the get_coin_data() function to obtain dataframe and inserting that dataframe in the function.
# coin_data <- get_coin_data(pair)
# low_52_week(coin_data)
}
