% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CryptoData.R
\name{plot_coin_data}
\alias{plot_coin_data}
\title{Plot Cryptocurrency Data}
\usage{
plot_coin_data(coin_data)
}
\arguments{
\item{coin_data}{Dataframe produced by the "get_coin_data" function.}
}
\description{
Plots cryptocurrency data using dataframe obtained through "get_coin_data" function using ggplot2.
}
\examples{
## Using the get_coin_data() function to obtain dataframe and inserting that dataframe in the function.
# coin_data <- get_coin_data(pair)
# plot_coin_data(coin_data)
}
